requirements.txt
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-nested-routers==0.93.4
psycopg2==2.9.4
flake8==4.0.1
flake8-html==0.4.2